[Chore] TableView Header Cell 수정 #9
